requirements.txt
playwright>=1.40
beautifulsoup4>=4.12
lxml>=4.9
pydantic>=1.10,<2.0
openai>=1.0
python-dotenv>=1.0
PyYAML>=6.0
gspread>=3.11